+++ src/util/types/command/PartialCommandType.ts
@@ -1,0 +1,5 @@
+import { Button } from "interactions.ts";
+
+export interface PartialCommand {
+  components: Button[];
+}
